move functions to search model
